refactor: applying fixes pointed by alRex-U
