+++ src/main/java/com/alrex/parcool/extern/betterthirdperson/BetterThirdPersonManager.java
@@ -1,9 +1,15 @@
 package com.alrex.parcool.extern.betterthirdperson;
 
 import com.alrex.parcool.common.action.impl.Dodge;
+import com.alrex.parcool.extern.AdditionalMods;
 import com.alrex.parcool.extern.ModManager;
+import com.alrex.parcool.utilities.VectorUtil;
+
 import io.socol.betterthirdperson.BetterThirdPerson;
+import io.socol.betterthirdperson.api.CustomCameraManager;
+import io.socol.betterthirdperson.api.util.AngleUtils;
 import net.minecraft.client.Minecraft;
+import net.minecraft.world.phys.Vec3;
 
 public class BetterThirdPersonManager extends ModManager {
     public BetterThirdPersonManager() {
@@ -18,10 +24,5 @@ public class BetterThirdPersonManager extends ModManager {
         return isInstalled()
             && !Minecraft.getInstance().options.getCameraType().isFirstPerson()
             && BetterThirdPerson.getCameraManager().hasCustomCamera();
-    }
-
-    public void cancelTurn() {
-        var cameraManager = BetterThirdPerson.getCameraManager();
-        cameraManager.stopPlayerTurning();
     }
 }
